requirements.txt
numpy==1.21.6
pandas==1.3.5
pycaret==3.1.0
matplotlib==3.5.3
pyyaml==6.0.1
fastapi
uvicorn[standard]